backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
pinecone
langchain
langchain-community
langchain-text-splitters
langchain-openai
langchain-groq
python-multipart
sqlalchemy
python-dotenv
pypdf
tiktoken
